// src/model/product.ts
export class Product {
  _id: string;
  // nome_produto: string;
  name: string;
  // preco_produto: number;
  price: number;
  // largura_produto: number;
  width: number;
  // comprimento_produto: number;
  length: number;
  // gramatura_produto: number;
  grammage: number;
  peso_produto: number;
  frete_produto: number;
  valor_total_produto: number;
  camisas_produto: number;
  custo_produto: number;
  dt_atualizacao: Date;
} 